refactor: split c_api module
